完善用户认证审核功能。1、根据status筛选得到不同状态的用户认证申请 2、管理员审核用户申请

+++ src/main/java/cn/edu/ncst/car/dao/UmsAdminRoleRelationDao.java
@@ -17,5 +17,5 @@ public interface UmsAdminRoleRelationDao {
     /**
      * 获取用户所有权限(包括+-权限)
      */
-    List<AuthPermission> getPermissionList(@Param("adminId") Long adminId);
+    List<AuthPermission> getPermissionList(@Param("adminId") Integer adminId);
 }
