+++ src/main/java/cn/edu/ncst/car/mbg/model/AuthRolePermissionRelation.java
@@ -4,35 +4,35 @@ import io.swagger.annotations.ApiModelProperty;
 import java.io.Serializable;
 
 public class AuthRolePermissionRelation implements Serializable {
-    private Long id;
+    private Integer id;
 
-    private Long roleId;
+    private Integer roleId;
 
-    private Long permissionId;
+    private Integer permissionId;
 
     private static final long serialVersionUID = 1L;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Long roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
 
-    public Long getPermissionId() {
+    public Integer getPermissionId() {
         return permissionId;
     }
 
-    public void setPermissionId(Long permissionId) {
+    public void setPermissionId(Integer permissionId) {
         this.permissionId = permissionId;
     }
